+++ src/components/Sidebar.svelte
@@ -1,74 +1,10 @@
 <script lang="ts">
-	// TODO: Automate generating profile image (hash email by md5)
-	// https://gravatar.com/site/implement/images/
-	let profileImage = 'https://www.gravatar.com/avatar/ff482bc4abad770478cc7bd36cb490fd';
-	let size = 200;
+	import Profile from './Profile.svelte';
 </script>
 
 <div data-testid="sidebar" {...$$restProps}>
 	<aside>
-		<!-- Avatar -->
-		<div data-testid="profile" class="avatar ml-3 mt-3 mb-5 md:ml-2 md:mt-2 md:mb-6 w-full">
-			<div class="w-1/3 md:w-3/5 rounded-full ring-2 ring-primary bg-primary">
-				<!-- svelte-ignore a11y-img-redundant-alt -->
-				<!-- BUG: https://github.com/sveltejs/language-tools/issues/524 -->
-				<img
-					class="mask mask-circle"
-					src={`${profileImage}?s=${size}`}
-					alt="My gravatar profile image"
-				/>
-			</div>
-		</div>
-
-		<!-- Introduction -->
-		<h1 class="font-semibold text-3xl md:text-2xl">Yuchan Lee</h1>
-		<p class="font-light text-xl md:text-md">A lazy developer.</p>
-
-		<!-- Contacts -->
-		<div class="mt-8 pr-4 flex justify-end">
-			<div class="flex space-x-3 rounded-lg p-2 px-3 dark:bg-white">
-				<!-- GitHub -->
-				<a href="https://github.com/lasuillard">
-					<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 98 96" class="w-9 h-9 md:w-7 md:h-7">
-						<path
-							fill-rule="evenodd"
-							clip-rule="evenodd"
-							d="M48.854 0C21.839 0 0 22 0 49.217c0 21.756 13.993 40.172 33.405 46.69 2.427.49 3.316-1.059 3.316-2.362 0-1.141-.08-5.052-.08-9.127-13.59 2.934-16.42-5.867-16.42-5.867-2.184-5.704-5.42-7.17-5.42-7.17-4.448-3.015.324-3.015.324-3.015 4.934.326 7.523 5.052 7.523 5.052 4.367 7.496 11.404 5.378 14.235 4.074.404-3.178 1.699-5.378 3.074-6.6-10.839-1.141-22.243-5.378-22.243-24.283 0-5.378 1.94-9.778 5.014-13.2-.485-1.222-2.184-6.275.486-13.038 0 0 4.125-1.304 13.426 5.052a46.97 46.97 0 0 1 12.214-1.63c4.125 0 8.33.571 12.213 1.63 9.302-6.356 13.427-5.052 13.427-5.052 2.67 6.763.97 11.816.485 13.038 3.155 3.422 5.015 7.822 5.015 13.2 0 18.905-11.404 23.06-22.324 24.283 1.78 1.548 3.316 4.481 3.316 9.126 0 6.6-.08 11.897-.08 13.526 0 1.304.89 2.853 3.316 2.364 19.412-6.52 33.405-24.935 33.405-46.691C97.707 22 75.788 0 48.854 0z"
-							fill="#24292f"
-						/>
-					</svg>
-				</a>
-
-				<!-- LinkedIn -->
-				<a href="https://www.linkedin.com/in/%EC%9C%A0%EC%B0%AC-%EC%9D%B4-67751020b/">
-					<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 72 72" class="w-9 h-9 md:w-7 md:h-7">
-						<path
-							d="M8,72 L64,72 C68.418278,72 72,68.418278 72,64 L72,8 C72,3.581722 68.418278,-8.11624501e-16 64,0 L8,0 C3.581722,8.11624501e-16 -5.41083001e-16,3.581722 0,8 L0,64 C5.41083001e-16,68.418278 3.581722,72 8,72 Z"
-							fill="#007EBB"
-						/>
-						<path
-							d="M62,62 L51.315625,62 L51.315625,43.8021149 C51.315625,38.8127542 49.4197917,36.0245323 45.4707031,36.0245323 C41.1746094,36.0245323 38.9300781,38.9261103 38.9300781,43.8021149 L38.9300781,62 L28.6333333,62 L28.6333333,27.3333333 L38.9300781,27.3333333 L38.9300781,32.0029283 C38.9300781,32.0029283 42.0260417,26.2742151 49.3825521,26.2742151 C56.7356771,26.2742151 62,30.7644705 62,40.051212 L62,62 Z M16.349349,22.7940133 C12.8420573,22.7940133 10,19.9296567 10,16.3970067 C10,12.8643566 12.8420573,10 16.349349,10 C19.8566406,10 22.6970052,12.8643566 22.6970052,16.3970067 C22.6970052,19.9296567 19.8566406,22.7940133 16.349349,22.7940133 Z M11.0325521,62 L21.769401,62 L21.769401,27.3333333 L11.0325521,27.3333333 L11.0325521,62 Z"
-							fill="#FFF"
-						/>
-					</svg>
-				</a>
-
-				<!-- Gmail -->
-				<a href="mailto:lasuillard@gmail.com">
-					<svg
-						xmlns="http://www.w3.org/2000/svg"
-						viewBox="52 42 88 66"
-						class="w-9 h-9 md:w-7 md:h-7"
-					>
-						<path fill="#4285f4" d="M58 108h14V74L52 59v43c0 3.32 2.69 6 6 6" />
-						<path fill="#34a853" d="M120 108h14c3.32 0 6-2.69 6-6V59l-20 15" />
-						<path fill="#fbbc04" d="M120 48v26l20-15v-8c0-7.42-8.47-11.65-14.4-7.2" />
-						<path fill="#ea4335" d="M72 74V48l24 18 24-18v26L96 92" />
-						<path fill="#c5221f" d="M52 51v8l20 15V48l-5.6-4.2c-5.94-4.45-14.4-.22-14.4 7.2" />
-					</svg>
-				</a>
-			</div>
-		</div>
+		<Profile class="ml-3 mt-3 mb-5 md:ml-2 md:mt-2 md:mb-6" />
 
 		<div class="divider" />
